add app insight env var to radix, and refactoring auth/config stuff
