+++ frontend/src/configuration.ts
@@ -4,6 +4,9 @@ interface Configuration {
     CLIENT_ID: string;
     TENANT_ID: string;
     APPINSIGHTS_CONNECTIONSTRING: string;
+    MSAL_SCOPES: Array<string>;
+    REDIRECT_URI: string;
+    AUTHORITY: string;
 }
 
 declare global {
@@ -21,12 +24,20 @@ function getEnvVars(): Configuration {
             CLIENT_ID: import.meta.env.VITE_CLIENT_ID,
             TENANT_ID: import.meta.env.VITE_TENANT_ID,
             APPINSIGHTS_CONNECTIONSTRING: import.meta.env.VITE_APPINSIGHTS_CONNECTIONSTRING,
+            MSAL_SCOPES: [import.meta.env.VITE_API_SCOPE],
+            REDIRECT_URI: window.location.origin,
+            AUTHORITY: "https://login.microsoftonline.com/3aa4a235-b6e2-48d5-9195-7fcf05b459b0",
         };
         return config;
     }
 
     // injectEnv is injected by inject-env.js based on inject-env-template.js
-    return window.injectEnv;
+    return {
+        ...window.injectEnv,
+        MSAL_SCOPES: [import.meta.env.VITE_API_SCOPE],
+        REDIRECT_URI: window.location.origin,
+        AUTHORITY: "https://login.microsoftonline.com/3aa4a235-b6e2-48d5-9195-7fcf05b459b0",
+    };
 }
 
 const config: Configuration = getEnvVars();
