+++ frontend/src/inject-env-template.js
@@ -4,4 +4,5 @@ window.injectEnv = {
     API_SCOPE: "${API_SCOPE}",
     CLIENT_ID: "${CLIENT_ID}",
     TENANT_ID: "${TENANT_ID}",
+    APPINSIGHTS_CONNECTIONSTRING: "${APPINSIGHTS_CONNECTIONSTRING}",
 };
