+++ frontend/src/services/auth.ts
@@ -1,24 +1,12 @@
-import {
-    AccountInfo,
-    AuthenticationResult,
-    BrowserCacheLocation,
-    EventType,
-    InteractionRequiredAuthError,
-    InteractionType,
-    PublicClientApplication,
-} from "@azure/msal-browser";
-import { AuthCodeMSALBrowserAuthenticationProvider } from "@microsoft/microsoft-graph-client/authProviders/authCodeMsalBrowser";
-
-import { config } from "../config/Settings";
-import { Client } from "@microsoft/microsoft-graph-client";
+import { BrowserCacheLocation, InteractionRequiredAuthError, PublicClientApplication } from "@azure/msal-browser";
 import { Providers, Msal2Provider } from "@microsoft/mgt";
-//import { Msal2Provider } from "@microsoft/mgt-msal2-provider";
+import config from "../configuration";
 
 // Initialize PublicClientApplication
 export const msalInstance: PublicClientApplication = new PublicClientApplication({
     auth: {
-        clientId: config.clientId || "",
-        authority: config.authority,
+        clientId: config.CLIENT_ID || "",
+        authority: config.AUTHORITY,
         redirectUri: window.location.origin,
     },
     cache: {
@@ -33,18 +21,10 @@ Providers.globalProvider = new Msal2Provider({
     scopes: ["User.Read", "People.Read", "User.ReadBasic.All"],
 });
 
-function getTenantAccount() {
-    const currentAccounts = msalInstance.getAllAccounts();
-    if (!currentAccounts) {
-        return null;
-    }
-    return currentAccounts.find((acc) => acc.tenantId === config.tenantId);
-}
-
 export async function getAccessToken(): Promise<string | null> {
     try {
         const tokenResponse = await msalInstance.acquireTokenSilent({
-            scopes: config.MsalScopes || [],
+            scopes: config.MSAL_SCOPES || [],
             forceRefresh: false,
         });
         return tokenResponse.accessToken;
@@ -52,7 +32,7 @@ export async function getAccessToken(): Promise<string | null> {
         console.error(error);
         if (error instanceof InteractionRequiredAuthError) {
             await msalInstance.acquireTokenRedirect({
-                scopes: config.MsalScopes || [],
+                scopes: config.MSAL_SCOPES || [],
             });
             return null;
         }
@@ -63,37 +43,3 @@ export async function getAccessToken(): Promise<string | null> {
 await msalInstance.initialize();
 msalInstance.enableAccountStorageEvents();
 await msalInstance.handleRedirectPromise();
-if (!msalInstance.getActiveAccount()) {
-    const account = getTenantAccount();
-    if (account !== undefined) {
-        msalInstance.setActiveAccount(account);
-    }
-}
-
-// msalInstance.addEventCallback((event) => {
-//     if (event.eventType === EventType.LOGIN_SUCCESS) {
-//         console.log("log in");
-//         const payload = event.payload as AuthenticationResult;
-//         msalInstance.setActiveAccount(payload.account);
-//     } else if (event.eventType === EventType.LOGOUT_SUCCESS) {
-//         console.log("log out");
-//         msalInstance.setActiveAccount(null);
-//     }
-// });
-
-async function testGraphApiAccess() {
-    const authProvider = new AuthCodeMSALBrowserAuthenticationProvider(msalInstance, {
-        account: msalInstance.getActiveAccount() as AccountInfo,
-        scopes: ["User.Read", "People.Read", "User.ReadBasic.All"],
-        interactionType: InteractionType.Popup,
-    });
-    const client = Client.initWithMiddleware({ authProvider });
-    try {
-        const me = await client.api("/me").get();
-        console.log("Graph API call successful:", me);
-    } catch (error) {
-        console.error("Graph API call failed:", error);
-    }
-}
-
-await testGraphApiAccess();
